Move the newly added test to an existing file
